fix(fe): change html tag
