+++ apps/frontend/app/(client)/(main)/contest/[contestId]/@tabs/leaderboard/_components/LeaderboardModalDialog.tsx
@@ -8,8 +8,62 @@ import {
 import infoIcon from '@/public/icons/file-info-gray.svg'
 import { X } from 'lucide-react'
 import Image from 'next/image'
-import { LeaderboardScoreExplanationRow } from './LeaderboardScoreExplanationRow'
-import { ModalNumberComponent } from './ModalNumberComponent'
+
+interface LeaderboardScoreExplanationRowProps {
+  isTalkBalloon: boolean
+  title: string
+  textColor?: string
+  explanationTitle: string
+  explanationDetail?: string
+}
+
+function LeaderboardScoreExplanationRow({
+  isTalkBalloon,
+  title,
+  textColor = '#000000',
+  explanationTitle,
+  explanationDetail
+}: LeaderboardScoreExplanationRowProps) {
+  return (
+    <div className="flex flex-row space-x-6">
+      {isTalkBalloon ? (
+        <div className="flex flex-col items-center justify-center">
+          <div className="relative mt-[9px] flex h-[33px] w-[70px] flex-row items-center justify-center rounded-full bg-[#3581FA] text-base font-semibold text-white">
+            {title}
+            <div className="absolute -bottom-[8px] left-1/2 h-0 w-0 -translate-x-1/2 border-l-[7px] border-r-[6px] border-t-[9px] border-l-transparent border-r-transparent border-t-[#3581FA]" />
+          </div>
+        </div>
+      ) : (
+        <p
+          className="mt-[9px] flex h-[50px] w-[70px] flex-row items-center justify-center rounded-[10px] bg-[#FAFAFA] text-xl font-semibold"
+          style={{ color: textColor }}
+        >
+          {title}
+        </p>
+      )}
+      <div className="flex h-[68px] w-[491px] flex-col justify-center rounded-[10px] bg-[#FAFAFA] pl-6">
+        <p className="h-[22px] w-[100px] font-semibold text-[#474747]">
+          {explanationTitle}
+        </p>
+        {explanationDetail && (
+          <p className="text-sm text-[#737373]">{explanationDetail}</p>
+        )}
+      </div>
+    </div>
+  )
+}
+
+interface ModalNumberComponentProps {
+  index: number
+}
+
+function ModalNumberComponent({ index }: ModalNumberComponentProps) {
+  return (
+    <div className="flex h-[22px] w-[22px] flex-col items-center justify-center rounded-full bg-[#3581FA] text-[11px] text-white">
+      {index}
+    </div>
+  )
+}
 
 export function LeaderboardModalDialog() {
   return (
@@ -38,56 +92,54 @@ export function LeaderboardModalDialog() {
             <p className="mb-8 h-[34px] w-[500px] text-[24px] font-bold text-black">
               Ranking System of Contest
             </p>
-            <p className="mb-7 flex text-[#5C5C5C]">
+            <div className="mb-7 flex text-[#5C5C5C]">
               <ModalNumberComponent index={1} />
               <div className="ml-3 flex flex-col">
-                <div className="text-base">
+                <p className="text-base">
                   The ranking of each participants is determined
-                </div>
-                <div className="text-base">
+                </p>
+                <p className="text-base">
                   based on the number of problems solved.
-                </div>
+                </p>
               </div>
-            </p>
-            <p className="flex text-[#5C5C5C]">
+            </div>
+            <div className="flex text-[#5C5C5C]">
               <ModalNumberComponent index={2} />
               <div className="ml-4 flex flex-col space-y-4">
                 <div>
-                  <div className="text-base">
+                  <p className="text-base">
                     If the number of problems solved is the same,
-                  </div>
-                  <div className="text-base">
+                  </p>
+                  <p className="text-base">
                     the ranking is determined in ascending order of penalty.
-                  </div>
+                  </p>
                 </div>
                 <div className="flex h-[104px] w-[562px] flex-col gap-3 rounded-[10px] bg-[#80808014] py-[18px] pl-6 pr-[69px] text-[#737373]">
-                  <div className="h-4 text-[14px] text-black">
+                  <p className="h-4 text-[14px] text-black">
                     <b>Penalty Calculation</b>
-                  </div>
+                  </p>
                   <div className="h-12">
-                    <div className="mb-1 text-[14px]">
+                    <p className="mb-1 text-[14px]">
                       Time from the start of the contest to Accepted (min.)
-                    </div>
-                    <div className="text-[14px]">
+                    </p>
+                    <p className="text-[14px]">
                       + (Number of submissions - 1) * 20
-                    </div>
+                    </p>
                   </div>
                 </div>
                 <div className="border-1 mt-7 flex h-[72px] w-[560px] flex-col justify-center rounded-[10px] border border-[#3581FA] pl-6 text-[#3581FA]">
-                  <div>
+                  <p>
                     If the number of problems solved and the penalty are the
                     same,
-                  </div>
-                  <div>they are considered to have the same rank.</div>
+                  </p>
+                  <p>they are considered to have the same rank.</p>
                 </div>
               </div>
-            </p>
-            <p className="pt-[76px]">
-              <div className="text-2xl font-bold">
-                Introduction to Leaderboard
-              </div>
-            </p>
-            <p className="space-y-4 pl-[11px] pt-[29px]">
+            </div>
+            <div className="pt-[76px]">
+              <p className="text-2xl font-bold">Introduction to Leaderboard</p>
+            </div>
+            <div className="space-y-4 pl-[11px] pt-[29px]">
               <LeaderboardScoreExplanationRow
                 isTalkBalloon={false}
                 title={'NNN'}
@@ -123,7 +175,7 @@ export function LeaderboardModalDialog() {
                   'The number of submissions displayed when hovering over ‘Accepted’'
                 }
               />
-            </p>
+            </div>
           </div>
         </DialogContent>
       </Dialog>
