+++ apps/frontend/app/(client)/(main)/contest/[contestId]/@tabs/leaderboard/_components/LeaderboardSolvedList.tsx
@@ -2,6 +2,7 @@ interface LeaderboardSolvedProps {
   solvedList: number[]
   problemSize: number
 }
+
 export function LeaderboardSolvedList({
   solvedList,
   problemSize
